requirements.txt
asttokens==3.0.1
beautifulsoup4==4.14.3
biopython==1.86
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
decorator==5.2.1
executing==2.2.1
Flask==3.1.2
flask-cors==6.0.1
fonttools==4.61.0
gget==0.29.3
idna==3.11
ipython==9.8.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jupyterlab_widgets==3.0.16
kiwisolver==1.4.9
lxml==6.0.2
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
mysql-connector-python==9.5.0
numpy==2.3.5
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pillow==12.0.0
prompt_toolkit==3.0.52
ptyprocess==0.7.0
pure_eval==0.2.3
py3Dmol==2.5.3
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
six==1.17.0
soupsieve==2.8
stack-data==0.6.3
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.1
wcwidth==0.2.14
Werkzeug==3.1.4
widgetsnbextension==4.0.15
